refactor(APP-1003): Rename graph util files with Utils suffix and colocate layout util
