--- apps/app/src/modules/settings/utils/permissionGraphLayout/index.ts
@@ -1,5 +1,0 @@
-export {
-    getLayoutedElements,
-    type IGetLayoutedElementsOptions,
-    type PermissionGraphDirection,
-} from './permissionGraphLayout';
